chore: skip wandb.Image tests if no pillow installed
