+++ requirements_test.txt
@@ -23,3 +23,5 @@ numpy~=1.24; python_version < '3.11'
 coverage[toml]~=7.2
 
 pyte~=0.8.1  # Terminal emulator for testing interactions with the terminal.
+
+hypothesis~=6.79  # Property-based testing for humans.
