+++ tests/unit_tests/test_media/test_media_logging.py
@@ -1,7 +1,3 @@
-import os
-import platform
-from pathlib import Path
-
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
@@ -23,11 +19,6 @@ def audio_media() -> wandb.Audio:
 def video_media() -> wandb.Video:
     frames = np.random.randint(low=0, high=256, size=(10, 3, 100, 100), dtype=np.uint8)
     return wandb.Video(frames)
-
-
-@pytest.fixture
-def image_media() -> wandb.Image:
-    return wandb.Image(np.ones(shape=(32, 32)))
 
 
 @pytest.fixture
@@ -85,7 +76,6 @@ def plotly_media() -> wandb.Plotly:
     "media_object",
     [
         "table_media",
-        "image_media",
         "video_media",
         "audio_media",
         "graph_media",
@@ -104,49 +94,3 @@ def test_log_media_saves_to_run_directory(mock_run, request, media_object):
 
     # Assert media object is saved under the run directory
     assert media_object._path.startswith(run.dir)
-
-
-@pytest.mark.parametrize(
-    "invalid_character",
-    [
-        "<",
-        ">",
-        ":",
-        "\\",
-        "?",
-        "*",
-    ],
-)
-@pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-def test_log_media_with_invalid_character_on_windows(
-    mock_run, image_media, invalid_character
-):
-    run = mock_run()
-    with pytest.raises(ValueError, match="Media .* is invalid"):
-        image_media.bind_to_run(run, f"image{invalid_character}test", 0)
-
-
-def test_log_media_with_path_traversal(mock_run, image_media):
-    run = mock_run()
-    image_media.bind_to_run(run, "../../../image", 0)
-
-    # Resolve to path to verify no path traversals
-    resolved_path = str(Path(image_media._path).resolve())
-    assert resolved_path.startswith(run.dir)
-    assert os.path.exists(resolved_path)
-
-
-@pytest.mark.parametrize(
-    "media_key",
-    [
-        "////image",
-        "my///image",
-    ],
-)
-def test_log_media_prefixed_with_multiple_slashes(mock_run, media_key, image_media):
-    run = mock_run()
-    image_media.bind_to_run(run, media_key, 0)
-
-    resolved_path = str(Path(image_media._path).resolve())
-    assert resolved_path.startswith(run.dir)
-    assert os.path.exists(resolved_path)
